Explicit LCMS linking for LibRAW on OSX and Linux
LCMS is needed for LibRAW. For some reason on OSX and Linux, this dependency needs to be explicitly found and linked.

I added LCMS to dependencies.txt.
On Linux, when I attempted to use vcpkg to download LibRAW, compiling caused it to link to my system install of LCMS instead.