Fixed two bugs found from testing.
The manual bit depth finder should only call get_bit_depth() on one of the white fields. This is because the white field should have the highest max pixel value. The artwork/target image may be less likely to use the full range of bits and the dark image very likely won't use the full range of bits either. So the manual bit depth finding algorithm may get an incorrect bit depth value from the dark image, and then use that value to over scale the artwork image and cause clipping.

The other bug was a simple mistake using a constant instead of the changing for loop value.