Added displayable names to refData dropdown options
Added fileName field to refData
Changed use of name field of refData to be the displayed name of the reference data.
Updated format of refData and interactions with refData to include fileName, standardObserver, and illuminates everywhere.
standardObserver and illuminates are effectively hardcoded on the frontend, however it should be much easier to implement setting both values in the front end in the future.
Added logic to update the value of the refData dropdown when selecting custom data.